* commit 'bb95452741bfd65723ea417c25dbd326c7568ab7':
  build: Fix ptr <-> integer cast warnings

+++ btif/src/btif_hh.c
@@ -434,7 +434,7 @@ void btif_hh_start_vup_timer(bt_bdaddr_t *bd_addr)
     {
         BTIF_TRACE_DEBUG("Start VUP timer ");
         memset(&p_dev->vup_timer, 0, sizeof(TIMER_LIST_ENT));
-        p_dev->vup_timer.param = (UINT32)btif_hh_tmr_hdlr;
+        p_dev->vup_timer.param = btif_hh_tmr_hdlr;
         btu_start_timer(&p_dev->vup_timer, BTU_TTYPE_USER_FUNC,
                         BTIF_TIMEOUT_VUP_SECS);
     }
